tidy gamestate

+++ src/Elite.Engine/Combat/Combat.cs
@@ -98,8 +98,8 @@ namespace Elite.Engine
 
                     if (ship_type is not ShipType.Planet and not ShipType.Sun)
                     {
-                        Space.universe[i].energy = _gameState.ShipList[(int)ship_type].energy;
-                        Space.universe[i].missiles = _gameState.ShipList[(int)ship_type].missiles;
+                        Space.universe[i].energy = _gameState.ShipList[ship_type].energy;
+                        Space.universe[i].missiles = _gameState.ShipList[ship_type].missiles;
                         Space.ship_count[ship_type]++;
                     }
 
@@ -441,9 +441,9 @@ namespace Elite.Engine
                 return;
             }
 
-            if (_gameState.ShipList[(int)type].scoopedType != StockType.None)
+            if (_gameState.ShipList[type].scoopedType != StockType.None)
             {
-                StockType trade = _gameState.ShipList[(int)type].scoopedType;
+                StockType trade = _gameState.ShipList[type].scoopedType;
                 _trade.AddCargo(trade);
                 _gameState.InfoMessage(_trade.stockMarket[(StockType)trade].name);
                 RemoveShip(un);
@@ -531,7 +531,7 @@ namespace Elite.Engine
             }
 
 
-            if (ship.energy < _gameState.ShipList[(int)type].energy)
+            if (ship.energy < _gameState.ShipList[type].energy)
             {
                 ship.energy++;
             }
@@ -599,7 +599,7 @@ namespace Elite.Engine
                 }
             }
 
-            maxeng = _gameState.ShipList[(int)type].energy;
+            maxeng = _gameState.ShipList[type].energy;
             energy = ship.energy;
 
             if (energy < (maxeng / 2))
@@ -632,7 +632,7 @@ namespace Elite.Engine
             direction = VectorMaths.VectorDotProduct(nvec, ship.Rotmat[2]);
 
             if ((ship.location.Length() < 8192) && (direction <= -0.833) &&
-                 (_gameState.ShipList[(int)type].laser_strength != 0))
+                 (_gameState.ShipList[type].laser_strength != 0))
             {
                 if (direction <= -0.917)
                 {
@@ -641,7 +641,7 @@ namespace Elite.Engine
 
                 if (direction <= -0.972)
                 {
-                    _ship.DamageShip(_gameState.ShipList[(int)type].laser_strength, ship.location.Z >= 0.0);
+                    _ship.DamageShip(_gameState.ShipList[type].laser_strength, ship.location.Z >= 0.0);
                     ship.acceleration--;
                     if (((ship.location.Z >= 0.0) && (_ship.ShieldFront == 0)) ||
                         ((ship.location.Z < 0.0) && (_ship.ShieldRear == 0)))
@@ -765,7 +765,7 @@ namespace Elite.Engine
                 return false;
             }
 
-            float size = _gameState.ShipList[(int)type].size;
+            float size = _gameState.ShipList[type].size;
 
             return ((x * x) + (y * y)) <= size;
         }
@@ -1100,7 +1100,7 @@ namespace Elite.Engine
                     return;
                 }
 
-                cnt &= _gameState.ShipList[(int)Space.universe[un].type].max_loot;
+                cnt &= _gameState.ShipList[Space.universe[un].type].max_loot;
                 cnt &= 15;
             }
